feat: update the message system to combine declaring a new worker and registration from a single class
