--- packages/fast-tooling-react/src/message-system-registry/index.ts
@@ -1,3 +1,0 @@
-import MessageSystemRegistry from "./message-system-registry";
-
-export { MessageSystemRegistry };
