+++ packages/fast-tooling-react/src/message-system/index.ts
@@ -1,9 +1,4 @@
-/// <reference lib="webworker" />
+import MessageSystem from "./message-system";
+import { MessageSystemType } from "./types";
 
-import { getMessage } from "./message-system";
-
-onmessage = function(e: MessageEvent): void {
-    if (e.data && e.data.type !== undefined) {
-        postMessage(getMessage(e.data));
-    }
-};
+export { MessageSystem, MessageSystemType };
