+++ packages/fast-tooling-react/src/message-system/types.ts
@@ -1,0 +1,7 @@
+export enum MessageSystemType {
+    data = "data",
+    dataDictionary = "data-dictionary",
+    navigation = "navigation",
+    navigationDictionary = "navigation-dictionary",
+    initialize = "initialize",
+}
