+++ packages/fast-tooling-react/src/message-system/webworker.ts
@@ -1,0 +1,9 @@
+/// <reference lib="webworker" />
+
+import { getMessage } from "./message-system.utilities";
+
+onmessage = function(e: MessageEvent): void {
+    if (e.data && e.data.type !== undefined) {
+        postMessage(getMessage(e.data));
+    }
+};
