--- packages/fast-tooling-react/tsconfig.build.message-system.json
@@ -1,18 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "target": "es6",
-        "outDir": "./dist/webworker",
-        "types": ["node"],
-        "lib": ["webworker", "es6"]
-    },
-    "include": [
-        "src/message-system/**/*"
-    ],
-    "exclude": [
-        "node_modules",
-        "**/*.spec.ts",
-        "**/*.spec.tsx",
-        "**/__tests__/**"
-    ]
-  }
